Added a new ways for a world to be generated, it can now be generated dynamically while also being contained to a maximum area defined in chunks

+++ src/main/java/inf/elte/hu/gameengine_javafx/Entities/TileEntity.java
@@ -36,7 +36,7 @@ public class TileEntity extends Entity {
 
         addToManager();
     }
-    public TileEntity(int value, int x, int y, String path) {
+    public TileEntity(int value, double x, double y, String path) {
         this.addComponent(new TileValueComponent(value));
         this.addComponent(new ImageComponent(path));
         this.addComponent(new PositionComponent(x, y, this));
